Extract nuisance estimation
